added SETVOL

+++ RoomServer/arduino/arduino.ino
@@ -59,7 +59,7 @@ void sendAudio(String c){
   Serial.println("Audio command: " + c);
   if (c.substring(0, 6) == "SETVOL"){
     IrSender.sendNEC(0xA002, "VOL-", 43);
-    IrSender.sendNEC(0xA002, "VOL+", c.substring(6, c.length()));
+    IrSender.sendNEC(0xA002, "VOL+", c.substring(6, c.length()).toInt());
   }
   else{
     IrSender.sendNEC(0xA002, comAudio(c), repeatAudio(c));
